ENH: Graph now contains the actual stream nodes. In case you wanted to play with the graph itself.
Potential use case:
Use a function to configure the task graph, then want to add things on.
We need to return a full representation of all the nodes, so you can
add to any one of them. The actual graph seems like a good way to do
that. Now if you want to add a sink to a node you just go to the graph,
get the stream_node for the associated node and run with that.