Register methods with Stream.register_api decorator
This moves methods off of the main stream class and offers an opportunity for
extension without subclassing.

People want to build their own operations for streams. The obvious way to do
this, subclassing, is hard due to how we have laid things out.  As an
alternative we consider function dispatch.  This also helps to reduce
boiler-plate, although it also adds some complexity and indirection.

Currently we define operations like `map` as a separate class and then add a
special method:

```python
class Stream(object):
    ...
    def map(...):
        return map(...

class map(Stream):
    ...
```

This is bad for two reasons:

1.  The map method is somewhat redundant
2.  We have hard coded all valid operations.  It's hard to add new ones.

As an alternative we dynamically add methods to stream with a decorator

```python
class Stream(object):
    @classmethod
    def register_api(cls, func):
        ...

@Stream.register_api
class map(Stream):
    ...
```