Add `from_end` kwarg to `Stream.from_textfile()`
Add a new keyword argument `from_end` to `Stream.from_textfile()` that
allows streaming from the end of the file. This allows for streams that
only emit lines that are appended to the file after `start()` is called.
Addresses #199.

Uses `file.seek(0, 2)` to seek to the end of the file before polling for
new lines.