Pick up dask's Client.loop when possible; fixes partition(timeout)
Originally, `Stream().partition(2, timeout=1.).scatter()` would result
in an error about two different event loops running. This is because the
`partition` command would require streamz to allocate an IOLoop, and
then later get the `dask.distributed.Client` event loop as well.

This patch introduces a guarded import in core to see if there is a dask
client running, and if so, uses that loop instead of starting a new one.

Also adds a regression test for async and sync behavior.