Allow for map_async to run items in parallel
Use an asyncio.Queue of the tasks to ensure that arrival and departure
order of elements match. Asserts back pressure when a new value arrives
via update but the work queue is full.

Because asyncio.Queue cannot peak, the parallelism factor is not
precise as the worker callback can have either zero or one task in hand
but it must free up a slot in the queue to do so. Under pressure, the
parallelism will generally be `(parallelism + 1)` instead of
`parallelism` as given in the `__init__` as one Future will be in the
awaited in the worker callback while the queue fills up from update
calls.