Source.from_tcp fixes
Running the network word count example showed an attempt to await a list
object in the Tornado framework. I tracked it down to the
EmitServer.handle_stream assuming the result of source._emit was
awaitable. Pivoting on inspect.isawaitable there makes the example as
written work.

I added a try/except block to trap the keyboard interrupt and suppress
the stack spew to the example.